backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
httpx==0.27.0
apscheduler==3.10.4
python-dotenv==1.0.1
